examples/mcp_image/requirements.txt
# Core framework dependency
mcp-agent @ file://../../  # Link to the local mcp-agent project root

# Additional dependencies specific to this example
anthropic
pillow